+++ tests/Functional/Entity/JobTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace App\Tests\Functional\Entity;
 
 use App\Entity\Job;
-use App\Tests\Services\InvokableFactory;
+use App\Tests\Services\InvokableFactory\JobSetupInvokableFactory;
 use App\Tests\Services\InvokableHandler;
 use webignition\SymfonyTestServiceInjectorTrait\TestClassServicePropertyInjectorTrait;
 
@@ -28,7 +28,7 @@ class JobTest extends AbstractEntityTest
      */
     public function testHydratedJobReturnsSourcesAsStringArray(array $sources)
     {
-        $this->invokableHandler->invoke(InvokableFactory::createJobWithSources(
+        $this->invokableHandler->invoke(JobSetupInvokableFactory::createJobWithSources(
             md5('label source'),
             'http://example.com/callback',
             $sources
